// src/lib/impakt-dev-api-client/react-query/types/getTimelineBlockRes.ts
/**
 * Generated by orval v6.10.2 🍺
 * Do not edit manually.
 * Impakt API
 * OpenAPI spec version: 1.0.0
 */
import type { GetTimelineBlockResType } from './getTimelineBlockResType';
import type { GetTimelineBlockResGroup } from './getTimelineBlockResGroup';
import type { GetTimelineBlockResExercise } from './getTimelineBlockResExercise';
import type { GetTimelineBlockAttributeRes } from './getTimelineBlockAttributeRes';

export interface GetTimelineBlockRes {
  id: number;
  order: number;
  type: GetTimelineBlockResType;
  group: GetTimelineBlockResGroup;
  /** null if timeline block type equal Rest */
  Exercise?: GetTimelineBlockResExercise;
  /** null if timeline block type equal Rest */
  exerciseId?: number | null;
  TimelineBlockAttributes: GetTimelineBlockAttributeRes[];
  createdAt: string;
  updatedAt: string;
  xpMax: number;
  godlMax: number;
  coinMax: number;
}